various enhancements for 52n-wps-io package. e.g. removed annoying System.out.println; changed use of "java.io.tmpdir" to File.createTempFile(); Created a pseudo generator for Shapefile-Directories. Can be used to overcome workflows like collect->zip->unzip.. This is also used within GTBinZippedSHPGenerator.
git-svn-id: https://svn.52north.org/svn/geoprocessing/main/WPS/trunk/WPS@2555 15ebfbcb-1bbe-4f7e-88aa-3b0fa5d666c2

+++ 52n-wps-io/src/main/java/org/n52/wps/io/IOUtils.java
@@ -153,16 +153,23 @@ public class IOUtils {
 	 *             if the unzipping process fails
 	 */
 	public static List<File> unzip(File file, String extension) throws IOException {
+		return unzip(file, extension, null);
+	}
+	
+	public static List<File> unzip(File file, String extension, File directory) throws IOException {
 		int bufferLength = 2048;
 		byte buffer[] = new byte[bufferLength];
 		List<File> foundFiles = new ArrayList<File>();
 		ZipInputStream zipInputStream = new ZipInputStream(
 				new BufferedInputStream(new FileInputStream(file)));
 		ZipEntry entry;
-		File tempDir = File.createTempFile("unzipped" + UUID.randomUUID(), "", new File(System
-				.getProperty("java.io.tmpdir")));
-		tempDir.delete();
-		tempDir.mkdir();
+		File tempDir = directory;
+		if (tempDir == null || !directory.isDirectory()) {
+			tempDir = File.createTempFile("unzipped" + UUID.randomUUID(), "", new File(System
+					.getProperty("java.io.tmpdir")));
+			tempDir.delete();
+			tempDir.mkdir();
+		}
 		while ((entry = zipInputStream.getNextEntry()) != null) {
 			int count;
 			File entryFile = new File(tempDir, entry.getName());
